Smarter timeouts.
Instead of checking all actions every minute, schedule a check when the closest one is expected to time out. Right now I just schedule the function that checks everything but it should be possible to schedule a function that gets passed the correct action etc, if one were to get really bored.